+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - change(internal): DirectionInterface 方法由 `parse` 改为 `guide`(#896)
 - change: 查询API方法由 `find` 改为 `query`，同时参数只支持 array(#897)
 - change: cancel 的 API 参数只支持 array，不再支持 string(#890)
+- change: close 的 API 参数只支持 array，不再支持 string(#891)
 
 ## v3.5.3